fed: Rename 'secret' controller package to 'sync'
