+++ federation/cmd/federation-controller-manager/app/BUILD
@@ -30,8 +30,8 @@ go_library(
         "//federation/pkg/federation-controller/ingress:go_default_library",
         "//federation/pkg/federation-controller/namespace:go_default_library",
         "//federation/pkg/federation-controller/replicaset:go_default_library",
-        "//federation/pkg/federation-controller/secret:go_default_library",
         "//federation/pkg/federation-controller/service:go_default_library",
+        "//federation/pkg/federation-controller/sync:go_default_library",
         "//pkg/util/configz:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
